+++ src/app/modules/reporters/components/index.ts
@@ -1,0 +1,1 @@
+export * from './ReportActionModal'
