+++ src/app/modules/responded/components/index.ts
@@ -1,0 +1,1 @@
+export * from './RespondTable'
