--- src/app/pages/DashboardWrapper.tsx
@@ -1,22 +1,0 @@
-/* eslint-disable jsx-a11y/anchor-is-valid */
-import {FC} from 'react'
-import {useIntl} from 'react-intl'
-import {PageTitle} from '../../_metronic/layout/core'
-
-const DashboardPage: FC = () => (
-  <>
-    <div></div>
-  </>
-)
-
-const DashboardWrapper: FC = () => {
-  const intl = useIntl()
-  return (
-    <>
-      <PageTitle breadcrumbs={[]}>{intl.formatMessage({id: 'MENU.DASHBOARD'})}</PageTitle>
-      <DashboardPage />
-    </>
-  )
-}
-
-export {DashboardWrapper}
